requirements.txt
aiofiles==23.2.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dill==0.3.8
dnspython==2.6.1
easyocr==1.7.1
email_validator==2.2.0
fastapi==0.111.1
fastapi-cli==0.0.4
ffmpy==0.3.2
filelock==3.15.4
fonttools==4.53.1
fsspec==2024.6.1
gradio==4.38.1
gradio_client==1.1.0
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.0
idna==3.7
imageio==2.34.2
importlib_resources==6.4.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.3
ninja==1.11.1.1
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
orjson==3.10.6
packaging==24.1
pandas==2.2.2
pillow==10.4.0
psutil==6.0.0
py-cpuinfo==9.0.0
pyclipper==1.3.0.post5
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.2
python-bidi==0.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.0
ruff==0.5.4
scikit-image==0.24.0
scipy==1.14.0
seaborn==0.13.2
semantic-version==2.10.0
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
sympy==1.13.1
thop==0.1.1.post2209072238
tifffile==2024.7.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.3.1
torchvision==0.18.1
tqdm==4.66.4
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.61
ultralytics-thop==2.0.0
urllib3==2.2.2
uvicorn==0.30.3
uvloop==0.19.0
watchfiles==0.22.0
websockets==11.0.3
